model: Handle stream updates in the compose header mid-compose.
This commit updates the compose header when a stream is updated while
the user is currently composing. The stream that the user is currently
writing a message to is checked and the relevant update is done if the
stream_id in the event matches the corresponding header.

Tests added.

Fixes #1040.