refactor: boxes/buttons/ui: Remove `emails` from private_box_view.
Prior to this commit, everytime the private_box_view had to be initiated
with recipients, the emails of the recipients was also requested along
with their user_ids. This commit removes that parameter and hence
makes the private_box_view reliant on only the list of `user_id`s
received.

Tests updated.