boxes: Use a helper method to set compose attributes to defaults.
This commit adds a helper method to the WriteBox class to set all
the compose-specific attributes to their defaults in the WriteBox's
`__init__` method as well as after usage by the `stream_box_view`
or `private_box_view`. This helps to ensure that the state of the
attributes do not persist beyond their function, because they are
specific to the compose action they were set in.

The `private_box_view` and `stream_box_view` set the `compose_box_view`
status appropriately in their call. The `GO_BACK` keypress now calls
this helper method to set all the compose attributes back to their
defaults.

Tests added.