bugfix: model: Remove updated `message_id`s from their old topic's set.
Prior to this commit, if a message's topic was changed, the
corresponding `message_id` was not removed from the old topic's set
under the "topic_msg_ids" key of the model's index. This commit removes
the updated `message_id` from that set, ensuring the accuracy of the
set.

Tests updated.